requirements.txt
asgiref==3.3.1
boto3==1.16.47
botocore==1.19.47
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
django-import-export==2.5.0
django-mathfilters==1.0.0
django-storages==1.11.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
